requirements.txt
asgiref==3.8.1
Django==3.2.20
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.1.1
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing-extensions==4.11.0
uritemplate==4.1.1
